+++ sciris/__init__.py
@@ -3,8 +3,8 @@ from .version import version as __version__, versiondate as __versiondate__, sci
 
 # Import core functions
 from .utils import *
-from .colortools import *
 from .odict import *
 from .dataframe import *
 from .fileio import *
 from .asd import *
+from .plotting import *
