+++ sciris/utils.py
@@ -2,33 +2,57 @@
 ### IMPORTS FROM OTHER LIBRARIES
 ##############################################################################
 
+
 import numpy as np
+import uuid as py_uuid
+import copy
+from textwrap import fill
+from pprint import pformat
+import datetime
+import dateutil
+from time import time, mktime, sleep
+from os import getcwd, remove
+from os.path import getsize
+if _PY2: from cPickle import dump
+else:    from pickle import dump
+from psutil import cpu_percent
+from subprocess import Popen, PIPE
+from functools import reduce
+from os import path, sep # Path and file separator
 
-
+# Imports required for slacknotification
+try:
+    from requests import post # Simple way of posting data to a URL
+    from json import dumps # For sanitizing the message
+    from getpass import getuser # In case username is left blank
+    _slackimports = 'success'
+except Exception as E:
+    _slackimports = repr(E)
+    
 # Handle types and Python 2/3 compatibility
 from six import PY2 as _PY2
 from numbers import Number as _numtype
 if _PY2: _stringtype = basestring 
 else:    _stringtype = str
 
+
 def uuid(uid=None, which=None, die=False, as_string=False):
     ''' Shortcut for creating a UUID; default is to create a UUID4. Can also convert a UUID. '''
-    import uuid
     if which is None: which = 4
-    if   which==1: uuid_func = uuid.uuid1
-    elif which==3: uuid_func = uuid.uuid3
-    elif which==4: uuid_func = uuid.uuid4
-    elif which==5: uuid_func = uuid.uuid5
+    if   which==1: uuid_func = py_uuid.uuid1
+    elif which==3: uuid_func = py_uuid.uuid3
+    elif which==4: uuid_func = py_uuid.uuid4
+    elif which==5: uuid_func = py_uuid.uuid5
     else: raise Exception('UUID type %i not recognized; must be 1,  3, 4 [default], or 5)' % which)
     
     if uid is None: # If not supplied, create a new UUID
         output = uuid_func()
     else: # Otherwise, try converting it
         try:
-            if isinstance(uid, uuid.UUID):
+            if isinstance(uid, py_uuid.UUID):
                 output = uid # Use directly
             else: # Convert
-                output = uuid.UUID(uid)
+                output = py_uuid.UUID(uid)
         except Exception as E: 
             errormsg = 'Could not covert "%s" to a UID (%s)' % (uid, repr(E))
             if die:
@@ -42,7 +66,6 @@ def uuid(uid=None, which=None, die=False, as_string=False):
 
 def dcp(obj=None):
     ''' Shortcut to perform a deep copy operation '''
-    import copy
     return copy.deepcopy(obj)
 
 
@@ -138,8 +161,8 @@ def objrepr(obj, showid=True, showmeth=True, showatt=True):
     return output
 
 
-def desc(obj, maxlen=None):
-    ''' Prints out the default representation of an object -- all attributes, plust methods and ID '''
+def prepr(obj, maxlen=None):
+    ''' Akin to "pretty print", returns a pretty representation of an object -- all attributes, plust methods and ID '''
     # Initialize things to print out
     if maxlen is None: maxlen = 80
     labels = []
@@ -165,8 +188,8 @@ def desc(obj, maxlen=None):
 
 
 def pr(obj, maxlen=None):
-    ''' Shortcut for printing the default repr for an object '''
-    print(desc(obj, maxlen=maxlen))
+    ''' Shortcut for printing the pretty repr for an object '''
+    print(prepr(obj, maxlen=maxlen))
     return None
 
     
@@ -191,10 +214,6 @@ def indent(prefix=None, text=None, suffix='\n', n=0, pretty=False, simple=True, 
     
     Version: 2017feb20
     '''
-    # Imports
-    from textwrap import fill
-    from pprint import pformat
-    
     # Handle no prefix
     if prefix is None: prefix = ' '*n
     
@@ -398,32 +417,11 @@ def printvars(localvars=None, varlist=None, label=None, divider=True, spaces=1, 
     return None
 
 
-def today(timezone='utc', die=False, tostring=False, fmt=None):
-    ''' Get the current time, in UTC time '''
-    import datetime # today = datetime.today
-    try:
-        import dateutil
-        if timezone=='utc':                           tzinfo = dateutil.tz.tzutc()
-        elif timezone is None or timezone=='current': tzinfo = None
-        else:                                         raise Exception('Timezone "%s" not understood' % timezone)
-    except Exception as E:
-        errormsg = 'Timezone information not available (%s)' % repr(E)
-        if die: raise Exception(errormsg)
-        tzinfo = None
-    now = datetime.datetime.now(tzinfo)
-    if tostring:
-        output = getdate(now)
-    else:
-        output = now
-    return output
 
-
-def getdate(obj=None, which='modified', fmt='str', dateformat=None):
+def getdate(obj=None, fmt='str', dateformat=None):
         ''' Return either the date created or modified ("which") as either a str or int ("fmt") '''
-        from time import mktime
-        
         if obj is None:
-            obj = today()
+            obj = now()
         
         if dateformat is None:
             dateformat = '%Y-%b-%d %H:%M:%S'
@@ -432,20 +430,24 @@ def getdate(obj=None, which='modified', fmt='str', dateformat=None):
             if isstring(obj): return obj # Return directly if it's a string
             obj.timetuple() # Try something that will only work if it's a date object
             dateobj = obj # Test passed: it's a date object
-        except: # It's not a date object
-            if which=='created': dateobj = obj.created
-            elif which=='modified': dateobj = obj.modified
-            elif which=='spreadsheet': dateobj = obj.spreadsheetdate
-            else: raise Exception('Getting date for "which=%s" not understood; must be "created", "modified", or "spreadsheet"' % which)
+        except Exception as E: # It's not a date object
+            raise Exception('Getting date failed; date must be a string or a date object: %s' % repr(E))
         
         if fmt=='str':
             try:
-                return dateobj.strftime(dateformat).encode('ascii', 'ignore') # Return string representation of time
+                output = dateobj.strftime(dateformat).encode('ascii', 'ignore') # Return string representation of time
+                return output
             except UnicodeDecodeError:
                 dateformat = '%Y-%m-%d %H:%M:%S'
-                return dateobj.strftime(dateformat)
-        elif fmt=='int': return mktime(dateobj.timetuple()) # So ugly!! But it works -- return integer representation of time
-        else: raise Exception('"fmt=%s" not understood; must be "str" or "int"' % fmt)
+                output = dateobj.strftime(dateformat)
+                return output
+        elif fmt=='int': 
+            output = mktime(dateobj.timetuple()) # So ugly!! But it works -- return integer representation of time
+            return output
+        else:
+            errormsg = '"fmt=%s" not understood; must be "str" or "int"' % fmt
+            raise Exception(errormsg)
+        return None # Should not be possible to get to this point
 
 
 
@@ -478,12 +480,6 @@ def slacknotification(to=None, message=None, fromuser=None, token=None, verbose=
         
     Version: 2017feb09
     '''
-    
-    # Imports
-    from requests import post # Simple way of posting data to a URL
-    from json import dumps # For sanitizing the message
-    from getpass import getuser # In case username is left blank
-    
     # Validate input arguments
     printv('Sending Slack message...', 2, verbose)
     if token is None:    token = '/.slackurl'
@@ -724,389 +720,7 @@ def promotetolist(obj=None, objtype=None, keepnone=False):
 
 
 
-##############################################################################
-### MATHEMATICAL FUNCTIONS
-##############################################################################
 
-
-def quantile(data, quantiles=[0.5, 0.25, 0.75]):
-    '''
-    Custom function for calculating quantiles most efficiently for a given dataset.
-        data = a list of arrays, or an array where he first dimension is to be sorted
-        quantiles = a list of floats >=0 and <=1
-    
-    Version: 2014nov23
-    '''
-    nsamples = len(data) # Number of samples in the dataset
-    indices = (np.array(quantiles)*(nsamples-1)).round().astype(int) # Calculate the indices to pull out
-    output = np.array(data)
-    output.sort(axis=0) # Do the actual sorting along the 
-    output = output[indices] # Trim down to the desired quantiles
-    
-    return output
-
-
-
-def sanitize(data=None, returninds=False, replacenans=None, die=True, defaultval=None, label=None, verbose=True):
-        '''
-        Sanitize input to remove NaNs. Warning, does not work on multidimensional data!!
-        
-        Examples:
-            sanitized,inds = sanitize(array([3,4,nan,8,2,nan,nan,nan,8]), returninds=True)
-            sanitized = sanitize(array([3,4,nan,8,2,nan,nan,nan,8]), replacenans=True)
-            sanitized = sanitize(array([3,4,nan,8,2,nan,nan,nan,8]), replacenans=0)
-        '''
-        try:
-            data = np.array(data,dtype=float) # Make sure it's an array of float type
-            inds = np.nonzero(~np.isnan(data))[0] # WARNING, nonzero returns tuple :(
-            sanitized = data[inds] # Trim data
-            if replacenans is not None:
-                newx = range(len(data)) # Create a new x array the size of the original array
-                if replacenans==True: replacenans = 'nearest'
-                if replacenans in ['nearest','linear']:
-                    sanitized = smoothinterp(newx, inds, sanitized, method=replacenans, smoothness=0) # Replace nans with interpolated values
-                else:
-                    naninds = inds = np.nonzero(np.isnan(data))[0]
-                    sanitized = dcp(data)
-                    sanitized[naninds] = replacenans
-            if len(sanitized)==0:
-                if defaultval is not None:
-                    sanitized = defaultval
-                else:
-                    sanitized = 0.0
-                    if verbose:
-                        if label is None: label = 'this parameter'
-                        print('sanitize(): no data entered for %s, assuming 0' % label)
-        except Exception as E:
-            if die: 
-                raise Exception('Sanitization failed on array: "%s":\n %s' % (repr(E), data))
-            else:
-                sanitized = data # Give up and just return an empty array
-                inds = []
-        if returninds: return sanitized, inds
-        else:          return sanitized
-
-
-def getvaliddata(data=None, filterdata=None, defaultind=0):
-    '''
-    Return the years that are valid based on the validity of the input data.
-    
-    Example:
-        getvaliddata(array([3,5,8,13]), array([2000, nan, nan, 2004])) # Returns array([3,13])
-    '''
-    data = np.array(data)
-    if filterdata is None: filterdata = data # So it can work on a single input -- more or less replicates sanitize() then
-    filterdata = np.array(filterdata)
-    if filterdata.dtype=='bool': validindices = filterdata # It's already boolean, so leave it as is
-    else:                        validindices = ~np.isnan(filterdata) # Else, assume it's nans that need to be removed
-    if validindices.any(): # There's at least one data point entered
-        if len(data)==len(validindices): # They're the same length: use for logical indexing
-            validdata = np.array(np.array(data)[validindices]) # Store each year
-        elif len(validindices)==1: # They're different lengths and it has length 1: it's an assumption
-            validdata = np.array([np.array(data)[defaultind]]) # Use the default index; usually either 0 (start) or -1 (end)
-        else:
-            raise Exception('Array sizes are mismatched: %i vs. %i' % (len(data), len(validindices)))    
-    else: 
-        validdata = np.array([]) # No valid data, return an empty array
-    return validdata
-
-
-
-def getvalidinds(data=None, filterdata=None):
-    '''
-    Return the years that are valid based on the validity of the input data from an arbitrary number
-    of 1-D vector inputs. Warning, closely related to getvaliddata()!
-    
-    Example:
-        getvalidinds([3,5,8,13], [2000, nan, nan, 2004]) # Returns array([0,3])
-    '''
-    data = promotetoarray(data)
-    if filterdata is None: filterdata = data # So it can work on a single input -- more or less replicates sanitize() then
-    filterdata = promotetoarray(filterdata)
-    if filterdata.dtype=='bool': filterindices = filterdata # It's already boolean, so leave it as is
-    else:                        filterindices = findinds(~np.isnan(filterdata)) # Else, assume it's nans that need to be removed
-    dataindices = findinds(~np.isnan(data)) # Also check validity of data
-    validindices = np.intersect1d(dataindices, filterindices)
-    return validindices # Only return indices -- WARNING, not consistent with sanitize()
-
-
-def approx(val1=None, val2=None, eps=None):
-    ''' Determine whether two scalars approximately match '''
-    if eps is None: eps = 1e-9
-    output = abs(val1-val2)<eps
-    return output
-
-
-def findinds(val1, val2=None, eps=1e-6):
-    '''
-    Little function to find matches even if two things aren't eactly equal (eg. 
-    due to floats vs. ints). If one argument, find nonzero values. With two arguments,
-    check for equality using eps. Returns a tuple of arrays if val1 is multidimensional,
-    else returns an array.
-    
-    Examples:
-        findinds(rand(10)<0.5) # e.g. array([2, 4, 5, 9])
-        findinds([2,3,6,3], 6) # e.g. array([2])
-    
-    Version: 2016jun06 
-    '''
-    if val2==None: # Check for equality
-        output = np.nonzero(val1) # If not, just check the truth condition
-    else:
-        if isstring(val2):
-            output = np.nonzero(np.array(val1)==val2)
-        else:
-            output = np.nonzero(abs(np.array(val1)-val2)<eps) # If absolute difference between the two values is less than a certain amount
-    if np.ndim(val1)==1: # Uni-dimensional
-        output = output[0] # Return an array rather than a tuple of arrays if one-dimensional
-    return output
-
-
-
-def findnearest(series=None, value=None):
-    '''
-    Return the index of the nearest match in series to value -- like findinds, but
-    always returns an object with the same type as value (i.e. findnearest with
-    a number returns a number, findnearest with an array returns an array).
-    
-    Examples:
-        findnearest(rand(10), 0.5) # returns whichever index is closest to 0.5
-        findnearest([2,3,6,3], 6) # returns 2
-        findnearest([2,3,6,3], 6) # returns 2
-        findnearest([0,2,4,6,8,10], [3, 4, 5]) # returns array([1, 2, 2])
-    
-    Version: 2017jan07
-    '''
-    series = promotetoarray(series)
-    if isnumber(value):
-        output = np.argmin(abs(series-value))
-    else:
-        output = []
-        for val in value: output.append(findnearest(series, val))
-        output = promotetoarray(output)
-    return output
-    
-    
-def dataindex(dataarray, index):        
-    ''' Take an array of data and return either the first or last (or some other) non-NaN entry. '''
-    
-    nrows = np.shape(dataarray)[0] # See how many rows need to be filled (either npops, nprogs, or 1).
-    output = np.zeros(nrows)       # Create structure
-    for r in range(nrows): 
-        output[r] = sanitize(dataarray[r])[index] # Return the specified index -- usually either the first [0] or last [-1]
-    
-    return output
-
-
-def smoothinterp(newx=None, origx=None, origy=None, smoothness=None, growth=None, ensurefinite=False, keepends=True, method='linear'):
-    '''
-    Smoothly interpolate over values and keep end points. Same format as numpy.interp.
-    
-    Example:
-        from utils import smoothinterp
-        origy = array([0,0.1,0.3,0.8,0.7,0.9,0.95,1])
-        origx = linspace(0,1,len(origy))
-        newx = linspace(0,1,5*len(origy))
-        newy = smoothinterp(newx, origx, origy, smoothness=5)
-        plot(newx,newy)
-        hold(True)
-        scatter(origx,origy)
-    
-    Version: 2018jan24
-    '''
-    # Ensure arrays and remove NaNs
-    if isnumber(newx):  newx = [newx] # Make sure it has dimension
-    if isnumber(origx): origx = [origx] # Make sure it has dimension
-    if isnumber(origy): origy = [origy] # Make sure it has dimension
-    newx  = np.array(newx)
-    origx = np.array(origx)
-    origy = np.array(origy)
-    
-    # If only a single element, just return it, without checking everything else
-    if len(origy)==1: 
-        newy = np.zeros(newx.shape)+origy[0]
-        return newy
-    
-    if not(newx.shape): raise Exception('To interpolate, must have at least one new x value to interpolate to')
-    if not(origx.shape): raise Exception('To interpolate, must have at least one original x value to interpolate to')
-    if not(origy.shape): raise Exception('To interpolate, must have at least one original y value to interpolate to')
-    if not(origx.shape==origy.shape): 
-        errormsg = 'To interpolate, original x and y vectors must be same length (x=%i, y=%i)' % (len(origx), len(origy))
-        raise Exception(errormsg)
-    
-    # Make sure it's in the correct order
-    correctorder = np.argsort(origx)
-    origx = origx[correctorder]
-    origy = origy[correctorder]
-    neworder = argsort(newx)
-    newx = newx[neworder] # And sort newx just in case
-    
-    # Only keep finite elements
-    finitey = np.isfinite(origy) # Boolean for whether it's finite
-    if finitey.any() and not finitey.all(): # If some but not all is finite, pull out indices that are
-        finiteorigy = origy[finitey]
-        finiteorigx = origx[finitey]
-    else: # Otherwise, just copy the original
-        finiteorigy = origy.copy()
-        finiteorigx = origx.copy()
-        
-    # Perform actual interpolation
-    if method=='linear':
-        newy = np.interp(newx, finiteorigx, finiteorigy) # Perform standard interpolation without infinities
-    elif method=='nearest':
-        newy = np.zeros(newx.shape) # Create the new array of the right size
-        for i,x in enumerate(newx): # Iterate over each point
-            xind = np.argmin(abs(finiteorigx-x)) # Find the nearest neighbor
-            newy[i] = finiteorigy[xind] # Copy it
-    else:
-        raise Exception('Method "%s" not found; methods are "linear" or "nearest"' % method)
-
-    # Perform smoothing
-    if smoothness is None: smoothness = np.ceil(len(newx)/len(origx)) # Calculate smoothness: this is consistent smoothing regardless of the size of the arrays
-    smoothness = int(smoothness) # Make sure it's an appropriate number
-    
-    if smoothness:
-        kernel = np.exp(-np.linspace(-2,2,2*smoothness+1)**2)
-        kernel /= kernel.sum()
-        validinds = findinds(~np.isnan(newy)) # Remove nans since these don't exactly smooth well
-        if len(validinds): # No point doing these steps if no non-nan values
-            validy = newy[validinds]
-            prepend = validy[0]*np.ones(smoothness)
-            postpend = validy[-1]*np.ones(smoothness)
-            if not keepends:
-                try: # Try to compute slope, but use original prepend if it doesn't work
-                    dyinitial = (validy[0]-validy[1])
-                    prepend = validy[0]*np.ones(smoothness) + dyinitial*np.arange(smoothness,0,-1)
-                except:
-                    pass
-                try: # Try to compute slope, but use original postpend if it doesn't work
-                    dyfinal = (validy[-1]-validy[-2])
-                    postpend = validy[-1]*np.ones(smoothness) + dyfinal*np.arange(1,smoothness+1,1)
-                except:
-                    pass
-            validy = np.concatenate([prepend, validy, postpend])
-            validy = np.convolve(validy, kernel, 'valid') # Smooth it out a bit
-            newy[validinds] = validy # Copy back into full vector
-    
-    # Apply growth if required
-    if growth is not None:
-        pastindices = findinds(newx<origx[0])
-        futureindices = findinds(newx>origx[-1])
-        if len(pastindices): # If there are past data points
-            firstpoint = pastindices[-1]+1
-            newy[pastindices] = newy[firstpoint] * np.exp((newx[pastindices]-newx[firstpoint])*growth) # Get last 'good' data point and apply inverse growth
-        if len(futureindices): # If there are past data points
-            lastpoint = futureindices[0]-1
-            newy[futureindices] = newy[lastpoint] * np.exp((newx[futureindices]-newx[lastpoint])*growth) # Get last 'good' data point and apply growth
-    
-    # Add infinities back in, if they exist
-    if any(~np.isfinite(origy)): # Infinities exist, need to add them back in manually since interp can only handle nan
-        if not ensurefinite: # If not ensuring all entries are finite, put nonfinite entries back in
-            orignan      = np.zeros(len(origy)) # Start from scratch
-            origplusinf  = np.zeros(len(origy)) # Start from scratch
-            origminusinf = np.zeros(len(origy)) # Start from scratch
-            orignan[np.isnan(origy)]     = np.nan  # Replace nan entries with nan
-            origplusinf[origy==np.inf]   = np.nan  # Replace plus infinite entries with nan
-            origminusinf[origy==-np.inf] = np.nan  # Replace minus infinite entries with nan
-            newnan      = np.interp(newx, origx, orignan) # Interpolate the nans
-            newplusinf  = np.interp(newx, origx, origplusinf) # ...again, for positive
-            newminusinf = np.interp(newx, origx, origminusinf) # ...and again, for negative
-            newy[np.isnan(newminusinf)] = -np.inf # Add minus infinity back in first
-            newy[np.isnan(newplusinf)]  = np.inf # Then, plus infinity
-            newy[np.isnan(newnan)]  = np.nan # Finally, the nans
-    
-    # Restore original sort order for newy
-    restoredorder = argsort(neworder)
-    newy = newy[restoredorder]
-    
-    return newy
-    
-
-def perturb(n=1, span=0.5, randseed=None):
-    ''' Define an array of numbers uniformly perturbed with a mean of 1. n = number of points; span = width of distribution on either side of 1.'''
-    if randseed is not None: np.random.seed(int(randseed)) # Optionally reset random seed
-    output = 1. + 2*span*(np.random.rand(n)-0.5)
-    return output
-    
-    
-def scaleratio(inarray,total):
-    ''' Multiply a list or array by some factor so that its sum is equal to the total. '''
-    origtotal = float(sum(inarray))
-    ratio = total/origtotal
-    outarray = np.array(inarray)*ratio
-    if type(inarray)==list: outarray = outarray.tolist() # Preserve type
-    return outarray
-
-
-def vec2obj(orig=None, newvec=None, inds=None):
-    ''' 
-    Function to convert an e.g. budget/coverage vector into an e.g. budget/coverage odict ...or anything, really
-    
-    WARNING: is all the error checking really necessary?
-    
-    inds can be a list of indexes, or a list of keys, but newvec must be a list, array, or odict.
-    
-    Version: 2016feb04
-    '''
-    # Validate input
-    if orig   is None: raise Exception('vec2obj() requires an original object to update')
-    if newvec is None: raise Exception('vec2obj() requires a vector as input')
-    lenorig = len(orig)
-    lennew  = len(newvec)
-    if lennew!=lenorig and inds is None: raise Exception('vec2obj(): if inds is not supplied, lengths must match (orig=%i, new=%i)' % (lenorig, lennew))
-    if inds is not None and max(inds)>=len(orig): 
-        raise Exception('vec2obj(): maximum index is greater than the length of the object (%i, %i)' % (max(inds), len(orig)))
-    if inds is None: inds = range(lennew)
-
-    # The actual meat of the function
-    new = dcp(orig)    
-    for i,ind in enumerate(inds):
-        new[ind] = newvec[i]
-    
-    return new
-
-
-def inclusiverange(*args, **kwargs):
-    '''
-    Like arange/linspace, but includes the start and stop points. 
-    Accepts 0-3 args, or the kwargs start, stop, step. Examples:
-    
-    x = inclusiverange(3,5,0.2)
-    x = inclusiverange(stop=5)
-    x = inclusiverange(6, step=2)
-    '''
-    
-    # Handle args
-    if len(args)==0:
-        start, stop, step = None, None, None
-    elif len(args)==1:
-        stop = args[0]
-        start, step = None
-    elif len(args)==2:
-        start = args[0]
-        stop   = args[1]
-        step = None
-    elif len(args)==3:
-        start = args[0]
-        stop = args[1]
-        step = args[2]
-    else:
-        raise Exception('Too many arguments supplied: inclusiverange() accepts 0-3 arguments')
-    
-    # Handle kwargs
-    start = kwargs.get('start', start)
-    stop  = kwargs.get('stop',  stop)
-    step  = kwargs.get('step',  step)
-    
-    # Finalize defaults
-    if start is None: start = 0
-    if stop  is None: stop  = 1
-    if step  is None: step  = 1
-    
-    # OK, actually generate
-    x = np.linspace(start, stop, int(round((stop-start)/float(step))+1)) # Can't use arange since handles floating point arithmetic badly, e.g. compare arange(2000, 2020, 0.2) with arange(2000, 2020.2, 0.2)
-    
-    return x
 
 
 
@@ -1115,6 +729,19 @@ def inclusiverange(*args, **kwargs):
 ### MISC. FUNCTIONS
 ##############################################################################
 
+def now(timezone='utc', die=False, tostring=False, fmt=None):
+    ''' Get the current time, in UTC time '''
+    if timezone=='utc':                           tzinfo = dateutil.tz.tzutc()
+    elif timezone is None or timezone=='current': tzinfo = None
+    else:                                         raise Exception('Timezone "%s" not understood' % timezone)
+    timenow = datetime.datetime.now(tzinfo)
+    if tostring:
+        output = getdate(timenow)
+    else:
+        output = timenow
+    return output
+    
+    
 
 def tic():
     '''
@@ -1123,7 +750,6 @@ def tic():
     toc() [but you can also use the form toc(t) where to is the output of tic()]
     '''
     global tictime  # The saved time is stored in this global.    
-    from time import time
     tictime = time()  # Store the present time in the global.
     return tictime    # Return the same stored number.
 
@@ -1135,8 +761,6 @@ def toc(start=None, output=False, label=None, sigfigs=None, filename=None):
     tic() [but you can also use the form t = tic()]
     toc() [but you can also use the form toc(t) where to is the output of tic()]
     '''   
-    from time import time
-    
     # Set defaults
     if label   is None: label = ''
     if sigfigs is None: sigfigs = 3
@@ -1181,11 +805,6 @@ def checkmem(origvariable, descend=0, order='n', plot=False, verbose=0):
         from utils import checkmem
         checkmem(['spiffy',rand(2483,589)],descend=1)
     '''
-    from os import getcwd, remove
-    from os.path import getsize
-    if _PY2: from cPickle import dump
-    else:    from pickle import dump
-    
     filename = getcwd()+'/checkmem.tmp'
     
     def dumpfile(variable):
@@ -1235,7 +854,8 @@ def checkmem(origvariable, descend=0, order='n', plot=False, verbose=0):
         print('Variable %s is %s' % (printnames[v], printsizes[v]))
 
     if plot==True:
-        from pylab import pie, array, axes
+        try:    from pylab import pie, array, axes
+        except: raise Exception('Cannot plot since import failed: %s' % repr(E))
         axes(aspect=1)
         pie(array(printbytes)[inds], labels=array(printnames)[inds], autopct='%0.2f')
 
@@ -1261,9 +881,6 @@ def loadbalancer(maxload=None, index=None, interval=None, maxtime=None, label=No
 
     Version: 2017oct25
      '''
-    from psutil import cpu_percent
-    from time import sleep
-    
     # Set up processes to start asynchronously
     if maxload  is None: maxload = 0.8
     if interval is None: interval = 5.0
@@ -1301,7 +918,6 @@ def runcommand(command, printinput=False, printoutput=False):
    
    Date: 2018mar28
    '''
-   from subprocess import Popen, PIPE
    if printinput: print(command)
    try:    output = Popen(command, shell=True, stdout=PIPE).communicate()[0].decode("utf-8")
    except: output = 'Shell command failed'
@@ -1312,7 +928,6 @@ def runcommand(command, printinput=False, printoutput=False):
 
 def gitinfo(filepath=None, die=False, hashlen=7):
     ''' Try to extract git information based on the file structure '''
-    from os import path, sep # Path and file separator
     if filepath is None: filepath = __file__
     try: # First try importing git-python
         import git
@@ -1383,7 +998,7 @@ def compareversions(version1=None, version2=None):
 ##############################################################################
 
 '''
-Four little functions to get and set data from nested dictionaries. The first two were stolen from:
+Four little functions to get and set data from nested dictionaries. The first two were adapted from:
     http://stackoverflow.com/questions/14692690/access-python-nested-dictionary-items-via-a-list-of-keys
 
 "getnested" will get the value for the given list of keys:
@@ -1428,7 +1043,6 @@ Version: 2014nov29
 
 def getnested(nesteddict, keylist, safe=False): 
     ''' Get a value from a nested dictionary'''
-    from functools import reduce
     output = reduce(lambda d, k: d.get(k) if d else None if safe else d[k], keylist, nesteddict)
     return output
 
@@ -1502,115 +1116,6 @@ def flattendict(inputdict=None, basekey=None, subkeys=None, complist=None, keyli
 
 
 
-##############################################################################
-### PLOTTING FUNCTIONS
-##############################################################################
-
-def setylim(data=None, ax=None):
-    '''
-    A small script to determine how the y limits should be set. Looks
-    at all data (a list of arrays) and computes the lower limit to
-    use, e.g.
-    
-        setylim([array([-3,4]), array([6,4,6])], ax)
-    
-    will keep Matplotlib's lower limit, since at least one data value
-    is below 0.
-    
-    Note, if you just want to set the lower limit, you can do that 
-    with this function via:
-        setylim(0, ax)
-    '''
-    # Get current limits
-    currlower, currupper = ax.get_ylim()
-    
-    # Calculate the lower limit based on all the data
-    lowerlim = 0
-    upperlim = 0
-    data = promotetolist(data) # Make sure it'siterable
-    for ydata in data:
-        lowerlim = min(lowerlim, promotetoarray(ydata).min())
-        upperlim = max(upperlim, promotetoarray(ydata).max())
-    
-    # Set the new y limits
-    if lowerlim<0: lowerlim = currlower # If and only if the data lower limit is negative, use the plotting lower limit
-    upperlim = max(upperlim, currupper) # Shouldn't be an issue, but just in case...
-    
-    # Specify the new limits and return
-    ax.set_ylim((lowerlim, upperlim))
-    return lowerlim,upperlim
-
-
-def SItickformatter(x, pos=None, sigfigs=2, SI=True, *args, **kwargs):  # formatter function takes tick label and tick position
-    ''' Formats axis ticks so that e.g. 34,243 becomes 34K '''
-    output = sigfig(x, sigfigs=sigfigs, SI=SI) # Pretty simple since sigfig() does all the work
-    return output
-
-
-def SIticks(fig=None, ax=None, axis='y', fixed=False):
-    ''' Apply SI tick formatting to one axis of a figure '''
-    from matplotlib import ticker
-    if  fig is not None: axlist = fig.axes
-    elif ax is not None: axlist = promotetolist(ax)
-    else: raise Exception('Must supply either figure or axes')
-    for ax in axlist:
-        if   axis=='x': thisaxis = ax.xaxis
-        elif axis=='y': thisaxis = ax.yaxis
-        elif axis=='z': thisaxis = ax.zaxis
-        else: raise Exception('Axis must be x, y, or z')
-        if fixed:
-            ticklocs = thisaxis.get_ticklocs()
-            ticklabels = []
-            for tickloc in ticklocs:
-                ticklabels.append(SItickformatter(tickloc))
-            thisaxis.set_major_formatter(ticker.FixedFormatter(ticklabels))
-        else:
-            thisaxis.set_major_formatter(ticker.FuncFormatter(SItickformatter))
-    return None
-
-
-def commaticks(fig=None, ax=None, axis='y'):
-    ''' Use commas in formatting the y axis of a figure -- see http://stackoverflow.com/questions/25973581/how-to-format-axis-number-format-to-thousands-with-a-comma-in-matplotlib '''
-    from matplotlib import ticker
-    if   ax  is not None: axlist = promotetolist(ax)
-    elif fig is not None: axlist = fig.axes
-    else: raise Exception('Must supply either figure or axes')
-    for ax in axlist:
-        if   axis=='x': thisaxis = ax.xaxis
-        elif axis=='y': thisaxis = ax.yaxis
-        elif axis=='z': thisaxis = ax.zaxis
-        else: raise Exception('Axis must be x, y, or z')
-        thisaxis.set_major_formatter(ticker.FuncFormatter(lambda x, p: format(int(x), ',')))
-    return None
-
-
-
-def boxoff(ax=None, removeticks=True, flipticks=True):
-    '''
-    I don't know why there isn't already a Matplotlib command for this.
-    
-    Removes the top and right borders of a plot. Also optionally removes
-    the tick marks, and flips the remaining ones outside.
-
-    Version: 2017may22    
-    '''
-    from pylab import gca
-    if ax is None: ax = gca()
-    ax.spines['top'].set_visible(False)
-    ax.spines['right'].set_visible(False)
-    if removeticks:
-        ax.xaxis.set_ticks_position('bottom')
-        ax.yaxis.set_ticks_position('left')
-    if flipticks:
-        ax.tick_params(direction='out', pad=5)
-    return ax
-
-
-
-
-        
-
-
 
 ##############################################################################
 ### LINKS
@@ -1643,7 +1148,7 @@ class Link(object):
     
     def __repr__(self):
         ''' Just use default '''
-        output  = desc(self)
+        output  = prepr(self)
         return output
     
     def __call__(self, obj=None):
